refactor(event_handler): 提取 MemoryRecall 模块
- 创建 core/event_handler_modules/memory_recall.py (337 行)
- 将记忆召回和注入逻辑提取到 MemoryRecall 类
- 包含记忆检索、注入方式选择、上下文扩展等完整逻辑
- 包含两个辅助方法：
  - _remove_injected_memories_from_context: 清理 system_prompt 和 extra_user_content_parts 中的旧记忆
  - _remove_fake_tool_call_from_context: 清理伪造的工具调用消息对
- EventHandler 中的 handle_memory_recall 简化为委托调用
- 更新测试以使用新的模块结构
- event_handler.py 从 1037 行减少到 572 行 (-465 行, -44.8%)
- 所有 44 个 event_handler 测试通过
