+++ core/event_handler.py
@@ -20,7 +20,7 @@ from .base.constants import (
     MEMORY_INJECTION_FOOTER,
     MEMORY_INJECTION_HEADER,
 )
-from .event_handler_modules import GroupCapture, MessageUtils
+from .event_handler_modules import GroupCapture, MemoryRecall, MessageUtils
 from .managers.conversation_manager import ConversationManager
 from .managers.memory_engine import MemoryEngine
 from .processors.memory_processor import MemoryProcessor
@@ -72,13 +72,21 @@ class EventHandler:
         self._group_capture = GroupCapture(
             config_manager, conversation_manager, self._message_utils
         )
+        self._injection_adapter = InjectionAdapter()
+        self._memory_recall = MemoryRecall(
+            context,
+            config_manager,
+            memory_engine,
+            conversation_manager,
+            self._message_utils,
+            self._injection_adapter,
+        )
 
         # 后台存储任务跟踪
         self._storage_tasks: set[asyncio.Task] = set()
         self._storage_sessions_inflight: set[str] = set()
         self._storage_state_lock = asyncio.Lock()
         self._shutting_down = False
-        self._injection_adapter = InjectionAdapter()
 
     async def handle_all_group_messages(self, event: AstrMessageEvent):
         """Capture all group messages for memory storage"""
@@ -86,238 +94,7 @@ class EventHandler:
 
     async def handle_memory_recall(self, event: AstrMessageEvent, req: ProviderRequest):
         """Query and inject long-term memory before LLM request"""
-        try:
-            session_id = event.unified_msg_origin
-            logger.debug(f"[DEBUG-Recall] 获取到 unified_msg_origin: {session_id}")
-
-            # 检测异常session_id
-            if session_id and (
-                "Error:" in session_id or "error:" in session_id.lower()
-            ):
-                logger.warning(
-                    f"[{session_id}] 检测到异常的session_id，这可能导致记忆功能异常。"
-                )
-
-            async with OperationContext("记忆召回", session_id):
-                prompt_text = getattr(req, "prompt", "")
-                extra_parts = getattr(req, "extra_user_content_parts", [])
-                has_prompt_text = isinstance(prompt_text, str) and bool(
-                    prompt_text.strip()
-                )
-                has_extra_parts = bool(extra_parts)
-
-                if not has_prompt_text and not has_extra_parts:
-                    logger.debug(f"[{session_id}] 请求中无可用用户内容，跳过记忆召回")
-                    return
-
-                # 自动删除旧的注入记忆
-                if self.config_manager.get("recall_engine.auto_remove_injected", True):
-                    removed = self._remove_injected_memories_from_context(
-                        req, session_id
-                    )
-                    removed += self._remove_fake_tool_call_from_context(req, session_id)
-                    if removed > 0:
-                        logger.info(
-                            f"[{session_id}] 已清理 {removed} 处历史记忆注入片段"
-                        )
-
-                # 先提取用户消息（消息存储和召回都需要）
-                actual_query = await self._message_utils.get_event_message_str(event)
-
-                request_query = (
-                    prompt_text.strip() if isinstance(prompt_text, str) else ""
-                )
-
-                # 存储用户消息（仅私聊），无论是否启用召回都需要
-                is_group = event.get_message_type() == MessageType.GROUP_MESSAGE
-                if not is_group and actual_query:
-                    message_to_store = request_query
-                    if not message_to_store:
-                        message_to_store = await self._message_utils.extract_message_content(
-                            event, req
-                        )
-                    if not message_to_store:
-                        message_to_store = actual_query.strip()
-                    await self.conversation_manager.add_message_from_event(
-                        event=event,
-                        role="user",
-                        content=message_to_store,
-                    )
-                    await self._message_utils.enforce_message_limit(session_id)
-
-                # 若 top_k <= 0，跳过记忆检索和注入，但上述清理和消息存储已执行
-                top_k = self.config_manager.get("recall_engine.top_k", 5)
-                if top_k <= 0:
-                    logger.info(
-                        f"[{session_id}] top_k={top_k} <= 0，跳过记忆检索和注入"
-                    )
-                    return
-
-                if not actual_query:
-                    logger.warning(f"[{session_id}] 原始用户消息为空，跳过记忆召回")
-                    return
-
-                # 获取过滤配置
-                filtering_config = self.config_manager.filtering_settings
-                use_persona_filtering = filtering_config.get(
-                    "use_persona_filtering", True
-                )
-                use_session_filtering = filtering_config.get(
-                    "use_session_filtering", True
-                )
-
-                # 获取 persona_id，与 AstrBot 主流程保持一致的三级优先级：
-                # 1. session_service_config（最高）
-                # 2. req.conversation.persona_id（会话级）
-                # 3. 全局默认人格（最低）
-                # 注意：on_llm_request 钩子在 _ensure_persona_and_skills 之前触发，
-                # 因此不能直接依赖 req.system_prompt 已注入人格，需自行走完整优先级。
-                persona_id = await get_persona_id(self.context, event)
-
-                recall_session_id = session_id if use_session_filtering else None
-                recall_persona_id = persona_id if use_persona_filtering else None
-
-                # 使用原始用户输入作为召回关键字
-                query_for_search = actual_query
-
-                # 上下文扩展：拼接最近2轮对话作为查询，提升检索精准度
-                if self.config_manager.get(
-                    "recall_engine.inject_with_recent_context", False
-                ):
-                    try:
-                        recent_messages = await self.conversation_manager.get_context(
-                            session_id, max_messages=5
-                        )
-                        if recent_messages and len(recent_messages) > 1:
-                            # recent_messages 按 timestamp DESC 排列（最新在前）
-                            # 跳过索引0（当前消息），取后续消息作为扩展上下文
-                            context_parts = []
-                            for msg in reversed(recent_messages[1:]):
-                                content = msg.get("content", "")
-                                if content and content.strip():
-                                    context_parts.append(content.strip())
-                            if context_parts:
-                                expanded = " | ".join(context_parts)
-                                query_for_search = expanded + " " + actual_query
-                                logger.info(
-                                    f"[{session_id}] 上下文扩展查询: "
-                                    f"{len(context_parts)}条历史消息 + 当前消息"
-                                )
-                    except Exception as e:
-                        logger.warning(f"[{session_id}] 获取上下文扩展失败: {e}")
-
-                # 执行记忆召回
-                logger.info(
-                    f"[{session_id}] 开始记忆召回，查询='{query_for_search[:80]}...'"
-                )
-
-                recalled_memories = await self.memory_engine.search_memories(
-                    query=query_for_search,
-                    k=self.config_manager.get("recall_engine.top_k", 5),
-                    session_id=recall_session_id,
-                    persona_id=recall_persona_id,
-                )
-
-                if recalled_memories:
-                    logger.info(
-                        f"[{session_id}] 检索到 {len(recalled_memories)} 条记忆"
-                    )
-
-                    # 格式化并注入记忆
-                    memory_list = [
-                        {
-                            "id": getattr(mem, "doc_id", None),
-                            "content": mem.content,
-                            "score": mem.final_score,
-                            "metadata": mem.metadata,
-                            "timestamp": mem.metadata.get("create_time"),
-                        }
-                        for mem in recalled_memories
-                    ]
-
-                    # 输出详细记忆信息
-                    for i, mem in enumerate(recalled_memories, 1):
-                        logger.debug(
-                            f"[{session_id}] 记忆 #{i}: 得分={mem.final_score:.3f}, "
-                            f"重要性={mem.metadata.get('importance', 0.5):.2f}, "
-                            f"内容={mem.content[:100]}..."
-                        )
-
-                    # 根据配置选择注入方式（含 Provider 兼容降级）
-                    configured_method = self.config_manager.get(
-                        "recall_engine.injection_method", "extra_user_content"
-                    )
-                    provider = None
-                    if configured_method == "fake_tool_call":
-                        provider = self.context.get_using_provider(session_id)
-                    injection_method, fallback_reason = self._injection_adapter.resolve(
-                        provider, configured_method
-                    )
-                    if fallback_reason:
-                        logger.warning(
-                            f"[{session_id}] 注入模式从 {configured_method} 降级为 "
-                            f"{injection_method}: {fallback_reason}"
-                        )
-
-                    memory_str = format_memories_for_injection(memory_list)
-
-                    if injection_method == "user_message_before":
-                        req.prompt = memory_str + "\n\n" + (req.prompt or "")
-                        logger.info(
-                            f"[{session_id}] 成功向用户消息前注入 {len(recalled_memories)} 条记忆"
-                        )
-                    elif injection_method == "user_message_after":
-                        req.prompt = (req.prompt or "") + "\n\n" + memory_str
-                        logger.info(
-                            f"[{session_id}] 成功向用户消息后注入 {len(recalled_memories)} 条记忆"
-                        )
-                    elif injection_method == "fake_tool_call":
-                        fake_messages = format_memories_for_fake_tool_call(
-                            memory_list,
-                            query=actual_query,
-                            k=self.config_manager.get("recall_engine.top_k", 5),
-                            session_filtered=use_session_filtering,
-                            persona_filtered=use_persona_filtering,
-                        )
-                        if fake_messages:
-                            req.contexts.extend(fake_messages)
-                            logger.info(
-                                f"[{session_id}] 成功以伪造工具调用方式注入 "
-                                f"{len(recalled_memories)} 条记忆"
-                            )
-                    elif injection_method == "fake_tool_call_deepseek_v4":
-                        fake_replay = format_memories_for_fake_tool_call_deepseek_v4(
-                            memory_list,
-                            query=actual_query,
-                            k=self.config_manager.get("recall_engine.top_k", 5),
-                            session_filtered=use_session_filtering,
-                            persona_filtered=use_persona_filtering,
-                        )
-                        if fake_replay:
-                            req.prompt = fake_replay + "\n\n" + (req.prompt or "")
-                            logger.info(
-                                f"[{session_id}] 成功以 DeepSeek V4 兼容伪工具转录方式注入 "
-                                f"{len(recalled_memories)} 条记忆"
-                            )
-                    else:
-                        # extra_user_content（推荐）：追加到用户消息末尾，
-                        # 不影响前缀缓存且 mark_as_temp 后不污染对话历史
-                        from astrbot.core.agent.message import TextPart
-
-                        req.extra_user_content_parts.append(
-                            TextPart(text=memory_str).mark_as_temp()
-                        )
-                        logger.info(
-                            f"[{session_id}] 成功向用户消息末尾注入 "
-                            f"{len(recalled_memories)} 条记忆"
-                        )
-                else:
-                    logger.info(f"[{session_id}] 未找到相关记忆")
-
-        except asyncio.CancelledError:
-            raise
-        except Exception as e:
-            logger.error(f"处理 on_llm_request 钩子时发生错误: {e}", exc_info=True)
+        await self._memory_recall.handle_memory_recall(event, req)
 
     async def handle_memory_reflection(
         self, event: AstrMessageEvent, resp: LLMResponse
@@ -772,248 +549,6 @@ class EventHandler:
             f"[{session_id}] 记录待重试总结: 范围=[{start_index}:{end_index}], "
             f"重试次数={new_retry_count}/3"
         )
-
-    def _remove_injected_memories_from_context(
-        self, req: ProviderRequest, session_id: str
-    ) -> int:
-        """从对话历史、system_prompt和prompt中删除之前注入的记忆片段"""
-
-        removed_count = 0
-        pattern = _INJECTION_CLEANUP_PATTERN
-
-        try:
-            # 清理 system_prompt（兼容旧版本注入残留）
-            if hasattr(req, "system_prompt") and req.system_prompt:
-                if isinstance(req.system_prompt, str):
-                    original_prompt = req.system_prompt
-                    if (
-                        MEMORY_INJECTION_HEADER in original_prompt
-                        and MEMORY_INJECTION_FOOTER in original_prompt
-                    ):
-                        cleaned_prompt = pattern.sub("", original_prompt)
-                        cleaned_prompt = re.sub(
-                            r"\n{3,}", "\n\n", cleaned_prompt
-                        ).strip()
-                        req.system_prompt = cleaned_prompt
-
-                        if cleaned_prompt != original_prompt:
-                            removed_count += 1
-                            logger.debug(
-                                f"[{session_id}] 从system_prompt中清理记忆片段 "
-                                f"(原长度={len(original_prompt)}, 新长度={len(cleaned_prompt)})"
-                            )
-
-            # 清理 extra_user_content_parts（处理 extra_user_content 注入方式）
-            if (
-                hasattr(req, "extra_user_content_parts")
-                and req.extra_user_content_parts
-            ):
-                kept_parts = []
-                for part in req.extra_user_content_parts:
-                    text = getattr(part, "text", "")
-                    if isinstance(text, str) and (
-                        MEMORY_INJECTION_HEADER in text
-                        and MEMORY_INJECTION_FOOTER in text
-                    ):
-                        removed_count += 1
-                        logger.debug(
-                            f"[{session_id}] 从extra_user_content_parts中清理记忆片段"
-                        )
-                        continue
-                    kept_parts.append(part)
-                req.extra_user_content_parts = kept_parts
-
-            # 清理 prompt（处理 user_message_before/after 注入方式）
-            if hasattr(req, "prompt") and req.prompt:
-                if isinstance(req.prompt, str):
-                    original_prompt = req.prompt
-                    if (
-                        MEMORY_INJECTION_HEADER in original_prompt
-                        and MEMORY_INJECTION_FOOTER in original_prompt
-                    ):
-                        cleaned_prompt = pattern.sub("", original_prompt)
-                        cleaned_prompt = re.sub(
-                            r"\n{3,}", "\n\n", cleaned_prompt
-                        ).strip()
-                        req.prompt = cleaned_prompt
-
-                        if cleaned_prompt != original_prompt:
-                            removed_count += 1
-                            logger.debug(
-                                f"[{session_id}] 从req.prompt中清理记忆片段 "
-                                f"(原长度={len(original_prompt)}, 新长度={len(cleaned_prompt)})"
-                            )
-
-            # 清理对话历史
-            if hasattr(req, "contexts") and req.contexts:
-                filtered_contexts = []
-
-                for _, msg in enumerate(req.contexts):
-                    # 处理三种格式:
-                    # 1. 字符串格式: "user: xxx"
-                    # 2. 字典+字符串内容: {"role": "user", "content": "xxx"}
-                    # 3. 字典+列表内容 (多模态): {"role": "user", "content": [{"type": "text", "text": "xxx"}]}
-
-                    if isinstance(msg, str):
-                        # 格式1: 字符串
-                        content = msg
-                    elif isinstance(msg, dict):
-                        content = msg.get("content", "")
-
-                        # 格式2和3: 字典
-                        if not isinstance(content, (str, list)):
-                            # 未知content类型,保留原消息
-                            filtered_contexts.append(msg)
-                            continue
-                    else:
-                        # 未知msg类型,保留原消息
-                        filtered_contexts.append(msg)
-                        continue
-
-                    # 处理字符串内容
-                    if isinstance(content, str):
-                        has_header = MEMORY_INJECTION_HEADER in content
-                        has_footer = MEMORY_INJECTION_FOOTER in content
-
-                        if has_header and has_footer:
-                            cleaned_content = pattern.sub("", content).strip()
-                            cleaned_content = re.sub(r"\n{3,}", "\n\n", cleaned_content)
-
-                            if not cleaned_content:
-                                removed_count += 1
-                                continue
-
-                            if cleaned_content != content:
-                                removed_count += 1
-                                if isinstance(msg, str):
-                                    filtered_contexts.append(cleaned_content)
-                                else:
-                                    msg_copy = msg.copy()
-                                    msg_copy["content"] = cleaned_content
-                                    filtered_contexts.append(msg_copy)
-                                continue
-
-                    # 处理列表内容 (多模态格式)
-                    elif isinstance(content, list):
-                        cleaned_parts = []
-                        has_changes = False
-
-                        for part in content:
-                            if isinstance(part, dict) and part.get("type") == "text":
-                                text = part.get("text", "")
-                                if isinstance(text, str):
-                                    has_header = MEMORY_INJECTION_HEADER in text
-                                    has_footer = MEMORY_INJECTION_FOOTER in text
-
-                                    if has_header and has_footer:
-                                        cleaned_text = pattern.sub("", text).strip()
-                                        cleaned_text = re.sub(
-                                            r"\n{3,}", "\n\n", cleaned_text
-                                        )
-
-                                        # 如果清理后为空,跳过这个part
-                                        if not cleaned_text:
-                                            has_changes = True
-                                            continue
-
-                                        # 如果清理后有内容,保留清理后的part
-                                        if cleaned_text != text:
-                                            has_changes = True
-                                            removed_count += 1
-                                            part_copy = part.copy()
-                                            part_copy["text"] = cleaned_text
-                                            cleaned_parts.append(part_copy)
-                                            continue
-
-                            cleaned_parts.append(part)
-
-                        # 如果整个content清理后为空,跳过整条消息
-                        if not cleaned_parts:
-                            removed_count += 1
-                            continue
-
-                        # 如果有修改,保存清理后的消息
-                        if has_changes:
-                            msg_copy = msg.copy()
-                            msg_copy["content"] = cleaned_parts
-                            filtered_contexts.append(msg_copy)
-                            continue
-
-                    # 未匹配到记忆标记,保留原消息
-                    filtered_contexts.append(msg)
-
-                req.contexts = filtered_contexts
-
-            if removed_count > 0:
-                logger.info(
-                    f"[{session_id}] 成功清理旧记忆片段，共删除 {removed_count} 处注入内容"
-                )
-
-        except Exception as e:
-            logger.error(f"[{session_id}] 删除注入记忆时发生错误: {e}", exc_info=True)
-
-        return removed_count
-
-    def _remove_fake_tool_call_from_context(
-        self, req: ProviderRequest, session_id: str
-    ) -> int:
-        """从对话历史中删除伪造的工具调用消息对。
-
-        识别并移除以 FAKE_TOOL_CALL_ID_PREFIX 为 ID 前缀的
-        assistant(tool_calls) + tool(result) 消息对。
-
-        Args:
-            req: LLM 请求对象。
-            session_id: 会话 ID，用于日志。
-
-        Returns:
-            删除的消息数量。
-        """
-
-        if not hasattr(req, "contexts") or not req.contexts:
-            return 0
-
-        removed = 0
-        indices_to_remove: set[int] = set()
-        fake_call_ids: set[str] = set()
-
-        try:
-            # 单轮扫描：同时收集伪造 assistant(tool_calls) 和对应 tool(result) 消息
-            # OpenAI 格式保证 assistant 在 tool 之前，因此 tool 匹配时 call_id 已就绪
-            for i, msg in enumerate(req.contexts):
-                if not isinstance(msg, dict):
-                    continue
-                role = msg.get("role")
-                if role == "assistant" and msg.get("tool_calls"):
-                    for tc in msg["tool_calls"]:
-                        tc_id = (
-                            tc.get("id", "")
-                            if isinstance(tc, dict)
-                            else getattr(tc, "id", "")
-                        )
-                        if tc_id.startswith(FAKE_TOOL_CALL_ID_PREFIX):
-                            fake_call_ids.add(tc_id)
-                            indices_to_remove.add(i)
-                elif role == "tool":
-                    tc_id = msg.get("tool_call_id", "")
-                    if tc_id in fake_call_ids:
-                        indices_to_remove.add(i)
-
-            # 从后往前删除，避免索引偏移
-            for i in sorted(indices_to_remove, reverse=True):
-                req.contexts.pop(i)
-                removed += 1
-
-            if removed > 0:
-                logger.info(f"[{session_id}] 清理了 {removed} 条伪造工具调用消息")
-
-        except Exception as e:
-            logger.error(
-                f"[{session_id}] 清理伪造工具调用时发生错误: {e}",
-                exc_info=True,
-            )
-
-        return removed
 
     async def handle_session_reset(self, event: AstrMessageEvent) -> None:
         """处理 /reset 或 /new 触发的会话清空，同步清除插件侧的消息历史和总结计数器"""
