+++ core/event_handler_modules/__init__.py
@@ -3,6 +3,7 @@ EventHandler 模块化子模块
 """
 
 from .group_capture import GroupCapture
+from .memory_recall import MemoryRecall
 from .message_utils import MessageUtils
 
-__all__ = ["MessageUtils", "GroupCapture"]
+__all__ = ["MessageUtils", "GroupCapture", "MemoryRecall"]
